update packages, add dry-run flag

+++ src/index.ts
@@ -10,20 +10,25 @@ import yargs from 'yargs';
 import SimpleGit from 'simple-git';
 
 const argv = yargs
-    .option("name", {
-        description: "The name",
-        alias: "n",
-        type: "string",
+    .usage("$0 <cmd> [args]")
+    .command("init", "Initialize a new node.js backend with express, Prisma, tsoa.")
+    .command("generate-test-route", "Generate a test route, so you can see that the setup is working.")
+    .command("generate-route <name> <route> <type>", "Generate a new API route", (yargs) => {
+        yargs.positional("name", {
+            type: "string",
+            describe: "the name of the newly generated controller for the route"
+        })
+        .positional("route", {
+            type: "string",
+            describe: "the API route to be generated"
+        })
+        .positional("type", {
+            type: "string",
+            describe: "\"Get\" or \"Post\""
+        })
     })
-    .option("route", {
-        alias: "r",
-        description: "The API route to be generated",
-        type: "string",
-    })
-    .option("type", {
-        alias: "t",
-        description: "Get or Post requst",
-        type: "string",
+    .option("dry-run", {
+        description: "run the generation without writing anything to disk",
     })
     .help()
     .alias("help", "h")
@@ -36,8 +41,7 @@ if (!fs.existsSync(projectPath + '/package.json')) {
 }
 
 if (argv._[0] === "init") {
-    console.log("Initializing project");
-    initProject();
+    initProject(argv.dryRun as boolean | undefined);
 } else if (argv._[0] === "generate-route") {
     if (!argv.name || !argv.route) {
         console.error("Need route for API and name for controller");
@@ -50,93 +54,109 @@ if (argv._[0] === "init") {
         type = "Post";
     }
     console.log("Generating code for route: ", argv.route);
-    generateApiRoute(argv.name, argv.route, type);
+    generateApiRoute(argv.name as string, argv.route as string, type, argv.dryRun as boolean | undefined);
 } else if (argv._[0] === "generate-test-route") {
-    generateApiRoute('test', 'test', 'Get');
+    generateApiRoute('test', 'test', 'Get', argv.dryRun as boolean | undefined);
 }
 
-function createFolderStructure(path: string, name: string) {
+function createFolderStructure(path: string, name: string, dryRun = false) {
     try {
-        fs.mkdirSync(path + '/src');
-        fs.mkdirSync(path + '/src/constants');
-        fs.mkdirSync(path + '/src/controllers');
-        fs.mkdirSync(path + '/src/engines');
-        fs.mkdirSync(path + '/src/services');
-        fs.mkdirSync(path + '/src/types');
-        fs.mkdirSync(path + '/build');
-        fs.mkdirSync(path + '/dist');
-        fs.mkdirSync(path + '/scripts');
-        fs.mkdirSync(path + '/static');
-        fs.mkdirSync(path + '/prisma');
+        if (!dryRun) {
+            fs.mkdirSync(path + '/src');
+            fs.mkdirSync(path + '/src/constants');
+            fs.mkdirSync(path + '/src/controllers');
+            fs.mkdirSync(path + '/src/engines');
+            fs.mkdirSync(path + '/src/services');
+            fs.mkdirSync(path + '/src/types');
+            fs.mkdirSync(path + '/build');
+            fs.mkdirSync(path + '/dist');
+            fs.mkdirSync(path + '/scripts');
+            fs.mkdirSync(path + '/static');
+            fs.mkdirSync(path + '/prisma');
+        }
+        console.log(`Created ${path}/src`);
+        console.log(`Created ${path}/src/constants`);
+        console.log(`Created ${path}/src/controllers`);
+        console.log(`Created ${path}/src/engines`);
+        console.log(`Created ${path}/src/services`);
+        console.log(`Created ${path}/src/types`);
+        console.log(`Created ${path}/build`);
+        console.log(`Created ${path}/dist`);
+        console.log(`Created ${path}/scripts`);
+        console.log(`Created ${path}/static`);
+        console.log(`Created ${path}/prisma`);
     } catch (e) {
         console.error(e);
     }
 }
 
-async function initProject() {
-    const port = GetRandomInt(8000, 9900);
+async function initProject(dryRun = false) {
     try {
+        console.log("\x1b[34mInitializing project\x1b[0m");
+        const port = GetRandomInt(8000, 9900);
 
         const packageJson = JSON.parse(fs.readFileSync(projectPath + '/package.json', { encoding: 'utf8' }));
         const name = packageJson.name;
-        createFolderStructure(projectPath, name);
+        createFolderStructure(projectPath, name, dryRun);
 
         if (packageJson["scripts"] === undefined) {
             packageJson["scripts"] = {};
         }
-        packageJson.scripts["build"] = "rimraf dist/* && tsoa spec-and-routes && tsc";
-        packageJson.scripts["dev"] = "npm run build && nodemon src/app.ts";
+        packageJson.scripts["clean"] = "rimraf dist";
+        packageJson.scripts["build"] = "tsoa spec-and-routes && tsc -b";
+        packageJson.scripts["dev"] = "npx nodemon --watch src/app.ts";
         packageJson.scripts["tsoa-gen"] = "tsoa routes && tsoa swagger";
         packageJson.scripts["tsoa-swagger"] = "tsoa swagger";
         packageJson.scripts["lint"] = "eslint . --ext .ts";
         packageJson.scripts["lint-and-fix"] = "eslint . --ext .ts --fix";
         packageJson.scripts["prettier-check"] = "npx prettier --check src";
         packageJson.scripts["prettier-fix"] = "npx prettier --write src";
-        packageJson.scripts["pm2"] = "pm2";
-        packageJson.scripts["start"] = "pm2 start cluster.json";
-        packageJson.scripts["stop"] = "pm2 kill";
-        packageJson.scripts["monit"] = "pm2 monit";
         packageJson.scripts["update-db-schema"] = "npx prisma db push && npx prisma generate";
 
         packageJson.main = "src/app.ts";
         if (packageJson["dependencies"] === undefined) {
             packageJson["dependencies"] = {};
         }
-        packageJson.dependencies["@prisma/client"] = "^4.5.0";
+        packageJson.dependencies["@prisma/client"] = "^5.20.0";
         packageJson.dependencies["compression"] = "^1.7.4";
         packageJson.dependencies["cors"] = "^2.8.5";
         packageJson.dependencies["dotenv"] = "^16.0.3";
         packageJson.dependencies["esm"] = "^3.2.25";
-        packageJson.dependencies["express"] = "^4.18.1";
-        packageJson.dependencies["helmet"] = "^4.2.0";
-        packageJson.dependencies["nodemon"] = "^2.0.19";
-        packageJson.dependencies["pm2"] = "5.2.0";
+        packageJson.dependencies["express"] = "^4.21.0";
+        packageJson.dependencies["helmet"] = "^8.0.0";
         packageJson.dependencies["rxjs"] = "^7.5.6";
-        packageJson.dependencies["swagger-ui-express"] = "^4.5.0";
-        packageJson.dependencies["ts-node"] = "^10.9.1";
-        packageJson.dependencies["tsoa"] = "^4.1.2";
-        packageJson.dependencies["typescript"] = "^4.8.3";
+        packageJson.dependencies["swagger-ui-express"] = "^5.0.1";
+        packageJson.dependencies["tsoa"] = "^6.4.0";
         packageJson.dependencies["typescript-ioc"] = "^3.2.2";
         if (packageJson["devDependencies"] === undefined) {
             packageJson["devDependencies"] = {};
         }
+
         packageJson.devDependencies["@types/compression"] = "^1.7.2";
         packageJson.devDependencies["@types/cors"] = "^2.8.12";
-        packageJson.devDependencies["@types/express"] = "^4.17.14";
-        packageJson.devDependencies["@types/node"] = "^18.7.18";
+        packageJson.devDependencies["@types/express"] = "^5.0.0";
+        packageJson.devDependencies["@types/node"] = "^22.7.4";
         packageJson.devDependencies["@types/swagger-ui-express"] = "^4.1.3";
-        packageJson.devDependencies["@typescript-eslint/eslint-plugin"] = "^5.37.0";
-        packageJson.devDependencies["@typescript-eslint/parser"] = "^5.37.0";
-        packageJson.devDependencies["eslint"] = "^8.23.1";
-        packageJson.devDependencies["eslint-config-prettier"] = "^8.5.0";
-        packageJson.devDependencies["eslint-plugin-prettier"] = "^4.0.0";
-        packageJson.devDependencies["prettier"] = "^2.7.1";
-        packageJson.devDependencies["prisma"] = "^4.5.0";
-        packageJson.devDependencies["rimraf"] = "^3.0.2";
+        packageJson.devDependencies["@typescript-eslint/eslint-plugin"] = "^8.8.0";
+        packageJson.devDependencies["@typescript-eslint/parser"] = "^8.8.0";
+        packageJson.devDependencies["eslint"] = "^8.57.1";
+        packageJson.devDependencies["eslint-config-prettier"] = "^9.1.0";
+        packageJson.devDependencies["eslint-plugin-prettier"] = "^5.2.1";
+        packageJson.devDependencies["nodemon"] = "^3.1.7";
+        packageJson.devDependencies["prettier"] = "^3.3.3";
+        packageJson.devDependencies["prisma"] = "^5.20.0";
+        packageJson.devDependencies["rimraf"] = "^6.0.1";
+        packageJson.devDependencies["typescript"] = "^5.6.2";
+        packageJson.devDependencies["ts-node"] = "^10.9.2";
         
-        fs.writeFileSync(projectPath + '/package.json', JSON.stringify(packageJson, null, 2));
+        if (!dryRun) {
+            fs.writeFileSync(projectPath + '/package.json', JSON.stringify(packageJson, null, 2));
+        }
+        console.log(`Created ${projectPath}/package.json`);
         console.log("Installing dependencies ....");
-        const childInstall = spawnSync('npm', ['install']);
+        if (!dryRun) {
+            const childInstall = spawnSync('npm', ['install']);
+        }
 
         const replaceName = new RegExp(/\$\{name\}/, 'g');
         const replacePort = new RegExp(/\$\{port\}/, 'g');
@@ -155,41 +175,64 @@ async function initProject() {
         const eslintignoreTemplate = fs.readFileSync(__dirname + "/../templates/.eslintignore.template", { encoding: 'utf8' }).replaceAll(replacePort, port.toString()).replaceAll(replaceName, name);
         const prettierignoreTemplate = fs.readFileSync(__dirname + "/../templates/.prettierignore.template", { encoding: 'utf8' }).replaceAll(replacePort, port.toString()).replaceAll(replaceName, name);
 
-        fs.writeFileSync(projectPath + '/tsconfig.json', tsconfigTemplate);
-        fs.writeFileSync(projectPath + '/tsoa.json', tsoaTemplate);
-        fs.writeFileSync(projectPath + '/src/app.ts', appTemplate);
-        fs.writeFileSync(projectPath + '/src/ioc.ts', iocTemplate);
-        fs.writeFileSync(projectPath + '/prisma/schema.prisma', prismaTemplate);
-        fs.writeFileSync(projectPath + '/.env', envTemplate);
-        fs.writeFileSync(projectPath + '/src/services/databaseService.ts', databaseServiceTemplate);
-        fs.writeFileSync(projectPath + '/README.md', readmeTemplate);
-        fs.writeFileSync(projectPath + '/cluster.json', clusterTemplate);
-        fs.writeFileSync(projectPath + '/.eslintrc.json', eslintrcTemplate);
-        fs.writeFileSync(projectPath + '/.prettierrc.json', prettierrcTemplate);
-        fs.writeFileSync(projectPath + '/.eslintignore', eslintignoreTemplate);
-        fs.writeFileSync(projectPath + '/.prettierignore', prettierignoreTemplate);
-
-        fs.openSync(projectPath + '/prisma/dev.db', 'w');
+        if (!dryRun) {
+            fs.writeFileSync(projectPath + '/tsconfig.json', tsconfigTemplate);
+            fs.writeFileSync(projectPath + '/tsoa.json', tsoaTemplate);
+            fs.writeFileSync(projectPath + '/src/app.ts', appTemplate);
+            fs.writeFileSync(projectPath + '/src/ioc.ts', iocTemplate);
+            fs.writeFileSync(projectPath + '/prisma/schema.prisma', prismaTemplate);
+            fs.writeFileSync(projectPath + '/.env', envTemplate);
+            fs.writeFileSync(projectPath + '/src/services/databaseService.ts', databaseServiceTemplate);
+            fs.writeFileSync(projectPath + '/README.md', readmeTemplate);
+            fs.writeFileSync(projectPath + '/cluster.json', clusterTemplate);
+            fs.writeFileSync(projectPath + '/.eslintrc.json', eslintrcTemplate);
+            fs.writeFileSync(projectPath + '/.prettierrc.json', prettierrcTemplate);
+            fs.writeFileSync(projectPath + '/.eslintignore', eslintignoreTemplate);
+            fs.writeFileSync(projectPath + '/.prettierignore', prettierignoreTemplate);
+            fs.openSync(projectPath + '/prisma/dev.db', 'w');
+        }
+        console.log(`Created ${projectPath}/tsconfig.json`);
+        console.log(`Created ${projectPath}/tsoa.json`);
+        console.log(`Created ${projectPath}/src/app.ts`);
+        console.log(`Created ${projectPath}/src/ioc.ts`);
+        console.log(`Created ${projectPath}/prisma/schema.prisma`);
+        console.log(`Created ${projectPath}/.env',`);
+        console.log(`Created ${projectPath}/src/services/databaseService.ts`);
+        console.log(`Created ${projectPath}/README.md`);
+        console.log(`Created ${projectPath}/cluster.json`);
+        console.log(`Created ${projectPath}/.eslintrc.json`);
+        console.log(`Created ${projectPath}/.prettierrc.json`);
+        console.log(`Created ${projectPath}/.eslintignore`);
+        console.log(`Created ${projectPath}/.prettierignore`);
+        console.log(`Created ${projectPath}/prisma/dev.db`);
 
         // Git stuff, check if folder already has a .git folder otherwise initialize git in the folder
         if (!fs.existsSync(projectPath + '/.git')) {
-            const git = SimpleGit(projectPath);
-            await git.init();
-            const gitignoreTemplate = fs.readFileSync(__dirname + "/../templates/.gitignore.template", { encoding: 'utf8' }).replaceAll(replacePort, port.toString());
-            fs.writeFileSync(projectPath + '/.gitignore', gitignoreTemplate);
-            await git.add(["package.json", "package-lock.json", "tsconfig.json", "tsoa.json", "src/", "scripts/", "prisma/", "static/", ".gitignore", ".eslintignore", ".eslintrc.json", ".prettierignore", ".prettierrc.json", "README.md", "cluster.json"]);
-            await git.commit(`Initializing project ${name}.\n\nUsing nodejs + express + tsoa + prisma`);
+            if (!dryRun) {
+                const git = SimpleGit(projectPath);
+                await git.init();
+                const gitignoreTemplate = fs.readFileSync(__dirname + "/../templates/.gitignore.template", { encoding: 'utf8' }).replaceAll(replacePort, port.toString());
+                fs.writeFileSync(projectPath + '/.gitignore', gitignoreTemplate);
+                await git.add(["package.json", "package-lock.json", "tsconfig.json", "tsoa.json", "src/", "scripts/", "prisma/", "static/", ".gitignore", ".eslintignore", ".eslintrc.json", ".prettierignore", ".prettierrc.json", "README.md", "cluster.json"]);
+                await git.commit(`Initializing project ${name}.\n\nUsing nodejs + express + tsoa + prisma`);
+            }
+            console.log("Initialized git");
+            console.log("Created .gitignore");
         }
 
+        console.log('\x1b[32mInitialization done!\x1b[0m You can generate a test API route with `npx gnb generate-test-route` or use `npx gnb generate-route <name> <route> <type>`.\nYou can run the project with `npm run dev` and it will be visible at http://localhost:' + port.toString() + '.');
+        if (dryRun) {
+            console.log('\x1b[43mThe command was run in dry-run mode, no files were written to disk.\x1b[0m');
+        }
     } catch (e) {
-        console.error(e);
+        console.error('Initialization failed\n', e);
     }
 
-    console.log('Done!');
 }
 
-function generateApiRoute(name: string, route: string, type: "Get" | "Post") {
+function generateApiRoute(name: string, route: string, type: "Get" | "Post", dryRun = false) {
     try {
+        console.log(`\x1b[34mGenerating new API route api/${name}/${route}\x1b[0m`);
         const replaceName = RegExp(/\$\{name\}/, 'g');
         const replaceRoute = RegExp(/\$\{route\}/, 'g');
         const replaceControllerName = RegExp(/\$\{controllerName\}/, 'g');
@@ -212,9 +255,16 @@ function generateApiRoute(name: string, route: string, type: "Get" | "Post") {
             .replaceAll(replaceName, name)    
             .replaceAll(replaceEngineName, capitalize(engineName));
 
-        fs.writeFileSync(projectPath + '/src/controllers/' + controllerName + '.ts', controllerTemplate);
-        fs.writeFileSync(projectPath + '/src/engines/' + engineName + '.ts', engineTemplate);
-        console.log('Done!');
+        if (!dryRun) {
+            fs.writeFileSync(projectPath + '/src/controllers/' + controllerName + '.ts', controllerTemplate);
+            fs.writeFileSync(projectPath + '/src/engines/' + engineName + '.ts', engineTemplate);
+        }
+        console.log(`Created ${projectPath}/src/controllers/${controllerName}.ts`);
+        console.log(`Created ${projectPath}/src/engines/${engineName}.ts`);
+        console.log(`Created new API route, available on BASE_URL/api/${name}/${route}.`);
+        if (dryRun) {
+            console.log('\x1b[43mThe command was run in dry-run mode, no files were written to disk.\x1b[0m');
+        }
     } catch (e) {
         console.error(e);
     }
